GOVSI-1113: Move `ApiGatewayHandlerIntegrationTest` to shared
- Move the base `ApiGatewayHandlerIntegrationTest` class to the `shared-test` module as this can be used in `account-management-integration-tests` too
