+++ integration-tests/src/test/java/uk/gov/di/authentication/api/JwksIntegrationTest.java
@@ -4,6 +4,7 @@ import com.nimbusds.jose.jwk.JWKSet;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import uk.gov.di.authentication.oidc.lambda.JwksHandler;
+import uk.gov.di.authentication.sharedtest.basetest.ApiGatewayHandlerIntegrationTest;
 
 import java.text.ParseException;
 import java.util.Map;
